first Version of the project
A main player have a Mazo and tile the Mazo have cards, he throw the cards in the table one by one. The players(as much as you want, In this case 4) battle one with another to win the cards. 
The cards have a type and a number.
While the match is in process, a rapporteur tell whats its happening in the process and save who is winning, By the end, the rapporteur say who win and with how many points. (the winner is the player that in his cards have the highest sum of numbers.)